add profiling data infrastructure
A new ProfileResult type is added to run_eval, which is
is returned in the EvalResult type. Among other fields,
this contains the `download_url` field which should be
used by the user to download profiling data. Note that
the actual public download link may not be known in
run_eval.py. In this case, it is the intention that the
launcher fixes up the `download_url` before returning the
results back to libkernelbot.